requirements.txt
# first install appropriate versions of cudnn and cudatoolkit for your hardware, i.e. run:
# conda install cudnn=8.2 cudatoolkit=11.3
# conda install --file requirements.txt -c pytorch

# Base ----------------------------------------
matplotlib>=3.2.2
numpy>=1.18.5
# opencv-python>=4.1.2
# Pillow>=7.1.2
PyYAML>=5.3.1
# requests>=2.23.0
scipy>=1.4.1
pytorch>=1.7.0
# torchvision>=0.8.1
tqdm>=4.41.0
nibabel>=3.3.3

# Logging -------------------------------------
# tensorboard>=2.4.1
# wandb

# Plotting ------------------------------------
pandas>=1.1.4
seaborn>=0.11.0

# Extras --------------------------------------
# thop  # FLOPs computation
